SONAR-6582 Extract issue serialization class
Use common issue JSON representation for most actions on issues.
